Mutex Lock Fix
Zephyr only supplies recursive mutexes. The way I implemented the
exclusive mutex didn't account for the fact that the mutex should wait
if it's on the same thread. Instead, it would just immediately return
false. Now, the lock will sleep for the set time then return false. This
assumes that other threads are not allowed to unlock the thread for the
thread that is trying to lock it. This case only happens when a
exclusive mutex is already locked by the same thread.