Fix lwip interface enumerate with single interface
Under lwip, os_enumerate_interfaces() would fail if only a single
interface is enabled by define.
In this situation NETIF_FOREACH define evaluates to an if statement,
which results in the interface count limit check attempting to exit this
if statement with a break.
As the limit check is redundant for a single interface, it is defined out.